+++ src/eConEXG/__init__.py
@@ -5,6 +5,7 @@ __all__ = [
     "iRecorder",
     "iFocus",
 ]
+from .version import __version__  # noqa: F401
 from .triggerBox import triggerBoxWired, triggerBoxWireless, lightStimulator
 from .iRecorder import iRecorder
 from .iFocus import iFocus