+++ src/eConEXG/iFocus/device_socket.py
@@ -2,12 +2,27 @@ import time
 
 
 class sock:
+    fs = {
+        250: b"\x04",
+        500: b"\x05",
+    }
+
     def __init__(self, port) -> None:
         from serial import Serial
 
         self.delay = 0.1
         self.dev = Serial(port=port, baudrate=921600, timeout=3)
+
+    def set_frequency(self, fs):
+        self.dev.flush()
         time.sleep(self.delay)
+        self.dev.write(sock.fs[fs])
+        time.sleep(self.delay)
+        res = self.dev.read_all()
+        if (fs == 500) and (len(res) == 0):
+            raise NotImplementedError(
+                "Invalid sample frequency, please update the device firmware or fall back to 250Hz."
+            )
 
     @staticmethod
     def _find_devs() -> list:
@@ -18,7 +33,7 @@ class sock:
         devices = comports()
         for device in devices:
             if "FTDI" in device.manufacturer:
-                if device.serial_number in ["IFOCUSA", "iFocus"]:
+                if "ifocus" in device.serial_number.lower():
                     try:
                         dev = Serial(port=device.device, baudrate=921600, timeout=1)
                     except serialutil.SerialException: